Massive changes in flags. Lots of deduplicated flags. More clean, unified.
Renamed `kvdb-loader` to `trxdb-loader`.

Clarified the meaning of `kvdb` (which is for connection strings), `trxdb`
is the name for the storage we use for blocks and transactions (and timeline
and accounts).. where `statedb` is what fluxdb stores (more about the state).
Things are symmetrical now as to how we specify database DSN.


Don't create `kvdb` and `fluxdb` in `dfuse-data` dir anymore.


See CHANGELOG.md