Greatly reduce memory allocation requires for DEOS -> EOS and EOS -> DEOS
All slices were allocated using `append` which for know length on input slice is known to
be less than ideal for re-allocation of the array. This was flagged by profiling console
reader work.

Now all arrays are pre-allocated, and an initial `len(in) <= 0` is also used to avoid
unecessary allocation of an empty array for cases where the input array is empty.