Fixed single table row read
We made the wrong assumption that the table index snapshot should contain the single row to
be read. But this is wrong. At the block height the snapshot was taken, it's quite possible
the row was not yet inserted.

We must not return row not found in this case. We must instead simply continue to normal
process and check the remaining rows, it might be in the next block segment that comes
after the index snapshot.

Fixed the error message also to now also include the account and table as well as using
the "string" form for the row instead of the hexadecimal form.

Fixes dfuse-io/customer-support#61