Fixed broken API behavior due to bug fix
When fixing the bug with `/state/table/row` since the old location of `DataRowNotFoundError`
was not the right place. So it was moved to a more appropriate location, but it changed
the behavior of the old API causing problem to existing customer. To retain old behavior,
we now return an empty row data in all cases when a specific key is not found on a given
table.