Fixed another slowness in `fluxdb` code
The limit value was not properly passed to `bigkv` and was not available at all for `Prefix`.

It's really important to pass a `limit` because the server can then send a single row. It's
a false impression that breaking after receiving a single result through the various layer
of `kvdb` abstraction that it's efficient. For example, BigTable without a limit set, returns
row in batch of let's say 100. We might receive multiple batches before reaching the context
cancel.

While setting a limit of 1 on BigTable forces the server to return at max 1 row which make
it much more efficient on both BigTable side and client side.