fix GetTransactionEvents to error if >1 trx match
When the idPrefix given was too short, events could be returned for more
than one transaction. This is dangerous if "merge" is called on those
events afterwards.

The `GetTransactionTraces` function and the `Batch` version of those two
calls will now return an error if a single prefix matches multiple
distinct trx IDs.

Tests were adapted because they were testing the wrong assumption.