Fixed GraphiQL not filling up predefined examples
Predefined examples should not properly be populated as long the `NetworkID` parameter of
dgraphql app is set correctly to `<protocol>-<network>` where `<protocol>` is `eos` and
the `<network>` is one of `mainnet`, `jungle` or `kylin`. If it's not one of those,
a `generic` value will be used by default

The predefined examples are not in `dfuse-eosio` directly instead of living in the
common `dgraphql` page.

Closes #143