Fixed race condition in eosws `get_transaction_lifecycle`
When the transaction was not found in the database, we were having an exclusive gate to
start at the block following the LIB. Then, a forkable handler was afterward setting the
LIB to be also exclusive, so starting also one block after LIB. But the last element
was an inclusive block id gate also on the LIB.

This condition was thus causing the handler to never start since the last inclusive LIB
would never seen the LIB block since it was excluded from previous handler.

We fixed the issue by making the forkable handler use and inclusive LIB, so the next handler
will correctly see it an pass through the block. We refactored at the same time the logic
so it's clearer the reasoning between the two gates and make the code more explicit.

This should fix Upland problem where some streams would never receive any messages.